# [Performance] Improve Infection performance executed against slow test suites