--- a/src/Command/RunCommand.php
+++ b/src/Command/RunCommand.php
@@ -116,6 +116,8 @@ final class RunCommand extends BaseCommand
 
     private const OPTION_USE_NOOP_MUTATORS = 'noop';
 
+    private const OPTION_EXECUTE_ONLY_COVERING_TEST_CASES = 'only-covering-test-cases';
+
     /** @var string */
     private const OPTION_MIN_MSI = 'min-msi';
 
@@ -256,6 +258,12 @@ final class RunCommand extends BaseCommand
                 null,
                 InputOption::VALUE_NONE,
                 'Use noop mutators that do not change AST. For debugging purposes.',
+            )
+            ->addOption(
+                self::OPTION_EXECUTE_ONLY_COVERING_TEST_CASES,
+                null,
+                InputOption::VALUE_NONE,
+                'Execute only those test cases that cover mutated line, not the whole file with covering test cases. Can dramatically speed up Mutation Testing for slow test suites. For PHPUnit / Pest it uses `--filter` option',
             )
             ->addOption(
                 self::OPTION_MIN_MSI,
@@ -444,7 +452,8 @@ final class RunCommand extends BaseCommand
             $gitDiffFilter,
             $gitDiffBase,
             (bool) $input->getOption(self::OPTION_LOGGER_GITHUB),
-            (bool) $input->getOption(self::OPTION_USE_NOOP_MUTATORS)
+            (bool) $input->getOption(self::OPTION_USE_NOOP_MUTATORS),
+            (bool) $input->getOption(self::OPTION_EXECUTE_ONLY_COVERING_TEST_CASES)
         );
     }