--- a/tests/phpunit/ContainerTest.php
+++ b/tests/phpunit/ContainerTest.php
@@ -95,7 +95,8 @@ final class ContainerTest extends TestCase
             Container::DEFAULT_GIT_DIFF_FILTER,
             Container::DEFAULT_GIT_DIFF_BASE,
             Container::DEFAULT_USE_GITHUB_LOGGER,
-            Container::DEFAULT_USE_NOOP_MUTATORS
+            Container::DEFAULT_USE_NOOP_MUTATORS,
+            Container::DEFAULT_EXECUTE_ONLY_COVERING_TEST_CASES
         );
 
         $traces = $newContainer->getUnionTraceProvider()->provideTraces();
@@ -144,7 +145,8 @@ final class ContainerTest extends TestCase
             Container::DEFAULT_GIT_DIFF_FILTER,
             Container::DEFAULT_GIT_DIFF_BASE,
             Container::DEFAULT_USE_GITHUB_LOGGER,
-            Container::DEFAULT_USE_NOOP_MUTATORS
+            Container::DEFAULT_USE_NOOP_MUTATORS,
+            Container::DEFAULT_EXECUTE_ONLY_COVERING_TEST_CASES,
         );
     }
 }