… tests/phpunit/TestFramework/PhpUnit/Adapter/PestAdapterTest.php
… tests/phpunit/TestFramework/PhpUnit/Adapter/PhpUnitAdapterTest.php
… tests/phpunit/TestFramework/PhpUnit/CommandLine/ArgumentsAndOptionsBuilderTest.php