This PR dramatically improves performance for Infection executed against functional (read - slow) tests.

```diff
- Time: 4m 18s. Memory: 0.05GB
+ Time: 1m 19s. Memory: 0.06GB
```

for one of the PR in my daily job's project.

## Problem

As you know, we always said that Infection runs only those tests that cover mutated line, not the all tests from an application.

It is partially true, but not exactly. PHPUnit's XML file allows us to only [filter original set of tests by `<directory>` or `<file>` ](https://phpunit.readthedocs.io/en/9.5/configuration.html#the-testsuite-element).

And this is what we do starting from 2017. 

<details>
<summary>Example of a `phpunit.xml` file generated by Infection for one of the Mutant (click me)</summary>


```xml
<?xml version="1.0" encoding="UTF-8"?>
<phpunit xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="./vendor/phpunit/phpunit/schema/9.2.xsd" bootstrap="/infection/tests/e2e/PestTestFramework/./infection/interceptor.autoload.740af9ed4589f3494f0c14ed544eec29.infection.php" colors="false" stopOnFailure="true" cacheResult="false" stderr="false">
  <testsuites>
    <testsuite name="Infection testsuite with filtered tests">
      <file>/infection/tests/e2e/PestTestFramework/tests/CalculatorPhpUnitTest.php</file>
    </testsuite>
  </testsuites>
  <filter>
    <whitelist>
      <directory>/infection/tests/e2e/PestTestFramework/src/</directory>
    </whitelist>
  </filter>
</phpunit>

```

</details>

Notice in the example above, that we tell PHPUnit to run tests only from `CalculatorPhpUnitTest.php`. But this file can contain 1 test that covers mutated line, **and hundreds of other tests** that still will be executed for Mutant. 

It wasn't always noticeable because we (ok, at least I am) used Infection for **unit** tests, which are fast enough.

However, since I'm currently working on a project with 2000+ *functional* tests and run Infection against this huge functional test suite, this became a bottleneck.

Imagine there is a file `ProductApiTest.php` that covers all the API endpoints with functional tests (Symfony WebTestCase).

* In this file, there is only 1 test case that covers mutated line. This test case takes **1s**
* In this file, there are additionally other 100 test cases that together take **15s**

So, with Infection 0.23.0 (current `master`), instead of running 1 test case for 1s, we run all the tests from `ProductApiTest.php` for 1+15=16s.

## Solution

With this PR, for the example explained above, we run only 1 test case for 1s, *saving 15 from 16 seconds*.

### Technical details

Since PHPUnit doesn't allow us to provide test cases on `phpunit.xml` file level, we can only do it on a command line level, using `--filter` option

* Command line on `master`: `'phpunit' '--configuration' '/path-to/phpunitConfiguration.d3b360aa95bfe11dadbe87dbd49b70c4.infection.xml'`
* Command line on this PR: `'phpunit' '--configuration' '/path-to/phpunitConfiguration.d3b360aa95bfe11dadbe87dbd49b70c4.infection.xml' '--filter' '/App\\Test::test_case1|App\\Test::test_case2/'` 

Instead of 1000 words, here is an example of the logs for the escaped mutant:

`master`:

```bash
3) /srv/api/src/DataPersister/TableViewDataPersister.php:32    [M] InstanceOf_

--- Original
+++ New
@@ @@
     }
     public function persist($data, array $context = [])
     {
-        if ($data instanceof TableView && ($context['collection_operation_name'] ?? null) === 'post' && $this->security->getUser() instanceof Employee) {
+        if ($data instanceof TableView && ($context['collection_operation_name'] ?? null) === 'post' && true) {
             /** @var Employee $employee */
             $employee = $this->security->getUser();
             $data->setCreatedBy($employee);

$ '/srv/api/vendor/phpunit/phpunit/phpunit' '--configuration' '/tmp/infection/phpunitConfiguration.1aadd4bc46a4b6ae51fb8e3a0d295e89.infection.xml'
  Test cache cleared
  PHPUnit 9.5.6 by Sebastian Bergmann and contributors.
  
  Random Seed:   1625689971
  
  Testing 
  .............................................................   61 / 1166 (  5%)
  .............................................................  122 / 1166 ( 10%)
  .............................................................  183 / 1166 ( 15%)
  .............................................................  244 / 1166 ( 20%)
  .............................................................  305 / 1166 ( 26%)
  .............................................................  366 / 1166 ( 31%)
  .............................................................  427 / 1166 ( 36%)
  .............................................................  488 / 1166 ( 41%)
  .............................................................  549 / 1166 ( 47%)
  .............................................................  610 / 1166 ( 52%)
  .............................................................  671 / 1166 ( 57%)
  ..............................
```

and this PR:

```bash
3) /srv/api/src/DataPersister/TableViewDataPersister.php:32    [M] InstanceOf_

--- Original
+++ New
@@ @@
     }
     public function persist($data, array $context = [])
     {
-        if ($data instanceof TableView && ($context['collection_operation_name'] ?? null) === 'post' && $this->security->getUser() instanceof Employee) {
+        if ($data instanceof TableView && ($context['collection_operation_name'] ?? null) === 'post' && true) {
             /** @var Employee $employee */
             $employee = $this->security->getUser();
             $data->setCreatedBy($employee);

$ '/srv/api/vendor/phpunit/phpunit/phpunit' '--configuration' '/tmp/infection/phpunitConfiguration.5b66e9f56127cb2a42531283cf8dc120.infection.xml' '--filter' '/App\\Tests\\[...very long filter]/'
  Test cache cleared
  PHPUnit 9.5.6 by Sebastian Bergmann and contributors.
  
  Random Seed:   1625689648
  
  Testing 
  ...............................................................  63 / 192 ( 32%)
  ............................................................... 126 / 192 ( 65%)
  ............................................................... 189 / 192 ( 98%)
  ...                                                             192 / 192 (100%)
```

So it is 192 executed tests (in this branch) instead of 1166 (in `master`). 

## How it impacts unit (fast) test suites

I tried to understand the impact for fast tests, running `master` branch and this PR against Infection itsetlf:

```bash
make test-infection-xdebug-80-docker
```

From my tests, running Infection against Infection becomes *slower* with this PR.

```
master:
  2m 52s. Memory: 0.17GB
  2m 47s. Memory: 0.17GB
  2m 46s. Memory: 0.17GB

PR:
  3m 8s. Memory: 0.19GB
  3m 6s. Memory: 0.19GB
  3m 9s. Memory: 0.19GB
```

* Why was memory usage increased (I don't think it's a problem though)? Because now we need to store all those `--filter '/the long list of test cases/' options
* Why was the execution time increased? From my tests, there are 2 reasons
  * Using `--filter` in PHPUnit itself adds some penalty as I understand, because this option is matched using `preg_match()` inside PHPUnit. But this adds only 5 extra seconds with the "Infection for Infection" example. 
  * Preparing a unique set of test cases and build `--filter '/the long list of test cases/' option takes the most extra time

Keeping that in mind, probably it's better to enable this feature by some option? If so, what the name of this new option you can recommend?

* `--only-covering-test-cases`
* `--slow-tests` (probably it will be some common option to enable not just this feature)
* ...

I don't know.

### Conclusion

The slower the tests, the more benefit / performance boost this PR will give.

---

- [x] Covered by tests
- [x] Doc PR: https://github.com/infection/site/pull/221