@sanmai 2021-07-19
Suggestion: if we can know a certain test suite contains this much tests (a thousand) which take this much seconds to run, we can make an educated guess as to whenever to enable this new option.