@maks-rafalko 2021-07-19
> Suggestion: if we can know a certain test suite contains this much tests (a thousand) which take this much seconds to run, we can make an educated guess as to whenever to enable this new option.

thousand of tests does not mean they are slow, to be honest. 1000 unit tests can take less than a second, while 1000 functional tests can take minutes. 

Probably we need to calculate the total time of the covering tests and compare with some constant. I tried it, and enabled `--filter` option dynamically, but my results was very strange or inaccurate. Probably will give it a one more try