@Ocramius 2021-07-20
@maks-rafalko I'm re-reading the description of this issue, and I'm trying to make sense of it.

Just because I'm trying to understand, if I have a test like following:

```php
class MyTest ... {
   function test1() { ... }
   function test2() { ... }
}
```

Does this patch make sure that only either of `test1` or `test2` are run for a specific mutation, if they cover separate code section? Did infection run all of `MyTest` before this patch?