@maks-rafalko 2021-07-20
> Does this patch make sure that only either of `test1` or `test2` are run for a specific mutation, if they cover separate code section? Did infection run all of `MyTest` before this patch?

@Ocramius  Exactly. Currently (in `master`), Infection creates a custom `phpunit.xml` for each mutant, and adds **test files** that cover mutated line, not particular test cases inside those files:


```xml
<?xml version="1.0" encoding="UTF-8"?>
<phpunit xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="./vendor/phpunit/phpunit/schema/9.2.xsd" bootstrap="/infection/tests/e2e/PestTestFramework/./infection/interceptor.autoload.740af9ed4589f3494f0c14ed544eec29.infection.php" colors="false" stopOnFailure="true" cacheResult="false" stderr="false">
  <testsuites>
    <testsuite name="Infection testsuite with filtered tests">
      <file>/infection/tests/e2e/PestTestFramework/tests/CalculatorPhpUnitTest.php</file>
    </testsuite>
  </testsuites>
  <filter>
    <whitelist>
      <directory>/infection/tests/e2e/PestTestFramework/src/</directory>
    </whitelist>
  </filter>
</phpunit>
```

^ as I described in the description, `CalculatorPhpUnitTest.php` test file can have 1000 tests cases, but only 10 cover mutated line. So, we need `--filter 'test_case_1|test_case_2...|test_case_10` to reduce the number of executed tests.