@Ocramius 2021-07-20
Excellent++! This is gonna be literal HOURS of mutation test runs saved. For me, the largest mutation test suite (integration tests) takes 2h :D