@maks-rafalko 2021-07-20
Yes, this is what I'm experiencing on my test suite as well with 2000+ functional tests ;)