@sanmai 2021-07-21
> I tried it, and enabled `--filter` option dynamically, but my results were very strange or inaccurate. Probably will give it one more try

Yes, it is all not that simple, unfortunately. I think this is worthy a dedicated research project, and an option toggle like `--only-covering-test-cases` should be the best course of action now. We can add an auto toggle later.