@maks-rafalko 2021-07-22
@sanmai ready for review. I've added `--only-covering-test-cases` option for now as we discussed above, so this new feature will be disabled by default and can be enabled manually