@nicojs 2021-08-02
Q: can the performance degradation on the "infection on infection" usecase also be influenced by simply having more tests in this branch? Or did you correct for that in some way?