@maks-rafalko 2021-08-02
I avoided the impact of mutating the new code and tests by adding `@infection-ignore-all` on the new code, and commenting out the tests. 

The most impact is really because of building the very long strings for each mutant where running all unit tests is faster than creating this filter